made the code modular and added a method for getting proxyy from proxylist site for piratebay as of now
