+++ config.py
@@ -1,0 +1,7 @@
+# Globally Declared Variables
+
+tor_seed = {}
+names = []
+seeds = []
+magnets = []
+sizes = []
